reverse last commit
originally the last commit was pushed to move users off the 3.x codebase , since very few are running the 9.3 codebase at this time and no one is mining with it i am pushing this commit to reverse the last commit. For anyone following the repo I did it this way vs. harder commands for anyone following the repo.